Serialized connecting to switch and fixed naming
I modified run-script to impose serialization in connecting to switch.
I used to get seg-faults before doing so, but never get seg-fault afterward.
This is also beneficial for simulating different network topologies as it
is necessary to know which gem5 process is connected to which port of
switch-box. In addition, I made some small changes to the script in order
to facilitate large scale simulation runs, by adding "rundir" and "ckptdir"
flags.

I also replaced "pd" and "multi" with "dist" extension.